간단하게 에코 서버 작성 및 클라이언트

+++ SimpleEcho/ServerConfig/ServerConfig.h
@@ -11,7 +11,7 @@ class ServerConfig {
 
   const uint16_t GetPort() const;
 
-  const uint16_t GetIp();
+  const uint32_t GetIp(); 
 
   const uint16_t GetThreadNo() const;
